+++ apps/demo/src/App.tsx
@@ -2,10 +2,11 @@ import { Routes, Route, BrowserRouter } from 'react-router-dom';
 
 import { Drawer } from './Drawer';
 import { Home } from './Home';
+import { YLSWrapper } from '@yourssu/logging-system-react';
 
 export const App = () => {
   return (
-    <>
+    <YLSWrapper baseURL={'https://7ybvghftyfxchg'}>
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<Home />} />
@@ -14,6 +15,6 @@ export const App = () => {
           </Route>
         </Routes>
       </BrowserRouter>
-    </>
+    </YLSWrapper>
   );
 };
